Add tap gesture to dismiss keyboard The desired behaviour is to dismiss the decimal keyboard when user taps anywhere outside the check amount field.
I added the .onTapGesture modifier to the Picker view and set the @FocusState property
to false, but as a side effect, picker doesn't list the options anymore.

Tried extending the View class with a function to force the first responder to resign
by sending an action to the shared application, but this didn't work either:
https://stackoverflow.com/questions/56491386/how-to-hide-keyboard-when-using-swiftui
https://www.hackingwithswift.com/quick-start/swiftui/how-to-dismiss-the-keyboard-for-a-textfield

All sources I found so far relies on developing complicated code, and full-screen white views.

I wonder if there's an built-in solution.